+++ spec/models/significant_date_history_reason_spec.rb
@@ -3,6 +3,18 @@ require "rails_helper"
 RSpec.describe SignificantDateHistoryReason do
   describe "Attributes" do
     it { is_expected.to have_db_column(:reason_type).of_type :string }
+
+    describe "reason_type" do
+      it "contains all the values used by the forms" do
+        earlier_values = DateHistory::Reasons::NewEarlierForm::REASONS_LIST
+        later_conversion_values = DateHistory::Reasons::NewLaterForm::CONVERSION_REASONS_LIST
+        later_transfer_values = DateHistory::Reasons::NewLaterForm::TRANSFER_REASONS_LIST
+
+        all_form_values = (earlier_values + later_transfer_values + later_conversion_values).map(&:to_s).uniq.sort
+
+        expect(all_form_values.to_set.subset?(SignificantDateHistoryReason.reason_types.keys.sort.to_set)).to be true
+      end
+    end
   end
 
   describe "Validations" do